pip-requirements.txt
nsenter
docker
python-consul
pyaml
